Create Sudoku class and implement __str__ method
This commit will create a new Sudoku.py which contains
a Sudoku class and an implementation of __str__ which
will display the board in a prettified way.
The Sudoku class requires a csv file to be initialized.
The csv file will need to contain the initial config of
the puzzle.